Partial implementation of EC2 DescribeInstanceStatus - EUCA-5052
Message binding and a partial implementation of DescribeInstanceStatus
are added for API versions from 2011-11-01. A new aws-instances binding
file adds bindings for the new request and response pair.
VmControl.groovy is updated to include the new message data objects.

VmControl implements the describe instance status operation and there is
a corresponding type mapper added to VmInstances. VmInstances is also
updated with filter support for the new operation.

Filtering is updated to allow the available filters to be qualified this
allows distinct filters to be available for describing instances and
describing instance status.

Filter and Filters are updated to allow combination of Filter items and
for use of a builder when constructing a Filter using internal filters.
An internal filter can be used for filtering on items that are not part
of the public filter API (e.g. by instance identifier for instance
status filtering)

EUCA-5052