Auto scaling EC2 reference validation - EUCA-5024
References to EC2 resources are now validated when auto scaling groups
and launch configurations are created/updated. The resources currently
validated are availability zones, image identifiers, ssh key names and
security group names.

The ActivityManager is updated with methods to validate references for
an owner and with a process task (ValidationScalingProcessTask) to
perform the work. There is an abstract activity task with sub-classes
for each resource type (ValidationScalingActivityTask). A Future is now
available to use if the results of a process task are required.

AutoScalingService is updated to use the new ActivityManager methods for
validation of referenced EC2 resources.

LaunchConfiguration and AutoScalingGroup are updated so collection
properties maintain their ordering.

EUCA-5024