Auto scaling instance lifecycle state support - EUCA-5190
Instances are now created in the Pending state and transitioned to
InService when a health check passes. Termination of an instance sets
the state to Terminating. If termination fails for any reason then the
operation will be retried until is succeeds (subject to the usual back
off)

LifecycleState is updated to provide a function to transition instance
states. PersistenceAutoScalingInstances exposes a new method to move
instances between states.

AbstractOwnedPersistents now provides a convenience method for
transactions with retries.

CollectionUtils is updated with some convenience methods for filtering
on object property values and evaluating predicates/functions for
iterables.

EUCA-5190