IAM roles crud and partial implementation of STS assume role - EUCA-5250
Support is added for IAM roles and a partial STS implementation of
issuing credentials for assuming roles.

Euare and token service bindings and PolicySpec are updated to support
the new operations. EuareService is updated with implementations for
role operation and stubs for instance profile operations.

Role, RoleEntity and DatabaseRoleProxy are added for role support.
Account and Privileged are updated with support for roles. RoleUser is
added for use when invoking operations with an assumed role.

TokensService and SecurityTokenManager are updated with a partial
implementation of the assume role operation. Roles can be assumed but
assume role policies are not currently supported/enforced.

Ern is updated to support ARNs for roles. RoleNumberQuotaKey is added to
support quotas for roles.

The EuareReplyQueue and EuareException are moved to the euare module for
improved modularity.

Account is updated to allow lookup of the admin user. Lookup of the
admin user by name is replaced with the new method where appropriate.

EUCA-5250