IAM / STS changes for assume role support - EUCA-5250
IAM policy support is added for policies attached to resources,
previously only policies attached to identities were supported. Resource
attached policies must specify a Princial/NotPrincipal. Policy parsing
now requires that the user specify the expected attachment type for
validation purposes.

Principal, PrincipalEntity, and DatabasePrincipalProxy are added for
principal support with policies. StatementEntity is updated to reference
the contained principal.

The policy engine and Permissions are updated to support evaluation of
resource attached policies and to allow inter-account resource denials.

A new RestrictedIdentity utility class is added along the lines of
RestrictedTypes but supporting resource attached policies.

PolicySpec now supports canonicalization of resource names which may or
may not be case sensitive depending on the vendor.

The DatabasePolicyProxy now support access to the authorizations for a
policy.

The TokensService is updated to enforce assume role policies.

Accounts now exposes utility methods for identity ARN construction.

EUCA-5250