Auto scaling load balancer add and remove instances - EUCA-5007
ActivityManager is updated to add and remove instances from the load
balancers registered with the group on launch and termination
respectively. New tasks are added for load balancer support and the main
scaling entry point is updated to perform retries for register and
deregister operations.

AutoScalingInstance now has a ConfigurationState which tracks the
current configuration level of each instance. Their is also a counter
for tracking registration and deregistration attempts.

CollectionUtils is updated to allow a reduction to minimum value for an
Integer collection.

EUCA-5007