Auto scaling zone rebalancing for configuration changes - EUCA-4995
Auto scaling availability zones are now rebalanced on configuration
change.

ActivityManager is updated to check if rebalancing is required whenever
a group requires scaling. The scaling required flag for a group is now
used to indicate rebalancing or capacity change is required.

AutoScaling group now exposes methods to update the desired capacity or
the availability zones to use. AutoScalingService uses these new methods
which ensure the scaling flag is set when necessary.

HealthStatus is updated to support use as a predicate for auto scaling
instances.

EUCA-4995