IAM instance profile support - EUCA-5250
Support is added for IAM instance profile CRUD including assignment of
roles to instance profiles.

An InstanceProfileEntity is added for persistence along with a related
InstanceProfile interface and DatabaseInstanceProfileProxy
implementation.

EuareService and Privileged are updated to support the instance profile
operations.

Roles and Accounts are updated to allow listing of related instance
profiles.

Instance profile related messages and bindings are updated to support
the custom account delegation parameter.

An InstanceProfileNumberQuotaKey is added for support of instance
profile quotas.

Prinipals is updated to clean up the implementation of system accounts.

EUCA-5250