Resource tagging updated to fix counting issue - EUCA-5360
There were a couple of issues with query by example for resource tag
counting. We were including the user information when we should have
been counting per account. The resource was not being included since it
was a separate entity (which was not joined)

A Criterion is now available for use with examples for correct matching.

A issue with updates failing when the tag limit was reached is also
addressed.

EUCA-5360