Auto scaling alarms in DescribePolicies response - EUCA-5465
Auto scaling now calls out to CloudWatch to determine the alarms that
are linked to scaling policies.

The AutoScalingService is updated to use ActivityManager for look up of
alarm ARNs. The service populates the response message with the alarm
information.

ActivityManager has new a process/activity for resovling alarm ARNs from
policy ARNs.

A CloudWatchClient is added for CloudWatch messages.

EUCA-5465