Impersonation is updated to unify external/internal mechanisms.
Impersonation is cleaned up to avoid duplicated impersonation code for
external and internal message paths.

Contexts now wraps the underlying Context when impersonation is active.
The impersonating context implements a new Wrapper interface that can be
used to extract the original Context if necessary. This is used by the
TagManager to verify that the caller has permission to modify restricted
tags.