EC2 AuthorizeSecurityGroupIngress updated for API / Group IDs- EUCA-3405
NetworkGroupManager is updated to support use of group IDs where
permitted in place of an account identifier / group name. NetworkGroups
exposes a method to resolve the ID for a group from the account
identifier / group name and vice versa. Group IDs are included in the
response when describing groups. NetworkPeer is updated to persist the
group ID in addition to the group name. NetworkRule is now initialized
by a list of NetworkPeers rather than a Map of account identifiers to
group names.

The query binding framework is updated to allow version information to
be specified for annotations. This allows the model to be annotated for
versions with conflicting naming requirements.

EUCA-3405