Auto scaling transaction cleanup - EUCA-5684
Auto scaling persistence is updated to avoid database access outside of
a transaction. View classes are introduced to represent data loaded from
entities. The immutable view classes eagerly load a subset of the
entities data, so use of a view guarantees that access will not fail due
to lazy loading.

EUCA-5684