Entities transaction cleanup - EUCA-5684
Entities is updated to close the EntityManager on commit or rollback of
a transaction. This change means that Hibernate entities using lazy
initialization will fail to load uninitialized aspects.

This change removes the possibility of database connections remaining
open for a long period of time after lazy initialization outside of any
transaction.

EUCA-5684