Hibernate 4 / JBoss Cache - Infinispan dependency update
Hibernate code and configuration is updated for version 4. JBoss Cache
is replaced with Infinispan. JGroups configuration is also updated.

Ehcache configuration files are removed.

VmMigrationTask is updated to remove a stray @Enumerated annotation.

CacheRegionFactory is responsible for creation of the CacheManager and
ensure that a single manager is created for all persistence contexts.
This prevents an error when Infinispan attempts to register the default
cache manager multiple times.

ConfigurableProxoolConnectionProvider is added to pass configuration
properties. This works around a Hibernate issue that prevents the
configuration being applied (HHH-7289)

Hosts is updated due to JGroups API changes.

EmpyreanTransactionManager now implements transaction manager lookup for
Infinispan.

Infinispan cache configuration is intended to be as similar to the JBoss
Cache configuration as possible. This is probably not a good
configuration as it will only cache data for one second.