EUCA-4738 Backup db creds on eucalyptus-cloud service start
Each time the eucalyptus-cloud service script is started it will check
to see if a backup of the current euca.p12 file exists. If not, it
will create a backup of it. If a previous backup exists it will warn
the user that their credentials have changed. It's up to the user to
restore the proper credentials from the backup location [1] if
necessary.

[1] $EUCALYPTUS/var/lib/eucalyptus/backups/keys