Changed DNS host matching for Walrus requests. Fixes EUCA-7666
Decouples DNS support in Walrus from the style of request being
sent and the inclusion of a service path.

With this change both path-style and hosted-style requests work
with DNS-enabled Walrus. Neither style requires a service path
(/services/Walrus) in the request. And if component-DNS is used
in the hostname of the request then any uri path will be treated
as a bucket.

For non-dns enabled clouds only path-style requests are accepted
and the service path must be included for proper request delivery