Support IAM service access using IAM role - EUCA-7870
Add TemporaryAccessKey extension of AccessKey for details of temporary
credentials such as the key type (role/session). The subjects public
QueryIdCredential now includes the key type to allow authorization
decisions to be based on this.

Services are updated to specify the key types they support and IAM now
allows role credentials.

EUCA-7870