Instance resource cleanup handling updates - EUCA-8499
Instance cleanup is broken down into two sections, one that is safe for
use on a non-persistent instance (entity). The order of cleanup tasks is
now explicit though use of a transaction synchronization ensuring the
initial transaction is committed prior to launch of the background
cleanup tasks, this ensure the entity updates cannot conflict.

EUCA-8499