EUCA-8707, EUCA-7938: Handle circumstances where listener blocks forever.
The Future's implementation of listener support failed to acknowledge that listeners are external code unaccompanied by durable guarantees about reliability. For example, they can choose to never return. Or return abnormally in some other way.  The calling side must address this.  To wit, the support for listeners in the futures now has a timeout.