Remove terminated instance cache - EUCA-8769
Terminated instances are now stored in the database rather than in an
in memory cache.

Persistent instances will be visible to users in the terminated state
until they expire. On expiry instances move to the buried state for a
period until being deleted.

There is now a configurable interval between updates to running instance
entities to fix and issue with detection of expired (unreported)
instances.

Resource tags for instances are now available when an instance is in the
terminated state, but the node tag will not be available when
terminated.

Redundant fields are added for instances to store copies of security
group and image identifiers and the instances virtualization type. This
allows references to other entities to be cleared on terminatation.

EUCA-8769