Error handling fixes for internal bindings and Auto Scaling - EUCA-8974
Error responses from services now set the return flag to false so they
can be detected as errors by the messaging framework. BindingHandler is
updated so it does not NPE when binding errors which was causing error
information to be lost.

Load balancing is updated to include the error code in error responses.

Auto scaling is updated to handle errors from the load Balancing service
and to respect the configured maximum retries for instance registration
and deregistration against a load balancer.

EUCA-8974