Mostly hardcode admin tool endpoint DNS names
Admin tools previously chose their endpoint URLs by replacing the
service path that appears in EC2_URL, which assumes that the EC2 service
is co-located with the control services.  Eucalyptus 4.0's new service
separation violates this assumption, so we deal with this by making each
instance in which EC2_URL would be consulted return http://127.0.0.1/
instead since we already require registration tools and the like to be
run from a CLC.

To change the endpoint's DNS name, use the -U/--url option, but note
that the service path will always be replaced with hardcoded values.
This is not a change from previous versions, but if it nonetheless
becomes an issue we can revisit it.

Fixes EUCA-8937