IP mapping for DNS service hosts - EUCA-9122
Added a cloud property to allow mapping of registered service host IPs
to alternate host IPs. This mapping allows an appropriate host IP to
returned for a given interface.

Mappings are specified as comma separated list of listener CIDR to host
CIDR:

dns.services.hostmapping = 172.16.55.0/24 -> 10.111.0.0/16

This example means the listener bound to a 172.16.55.10 address would
map a service IP to an address within 10.111.0.0/16 on the service host.

EUCA-9122