Make euca_conf capable of syncing euca.p12 for 4.0 upgrade
The way eucalyptus does service registration means that every Java-based
service host has to obtain a new copy of euca.p12 after a new component
type is added to the system.  This would normally require one to run
euca_conf --synckey for each machine running such a service, so we
streamline this a bit by adding a new --sync-euca-p12 option that copies
euca.p12 to all machines (except for the enabled CLC, which this should
be running from) that are running at least one of the following four
services after an upgrade to eucalyptus 4.0:
  * bootstrap
  * storage
  * vmwarebroker
  * walrusbackend

This operation is very selective about what it copies and where it
copies to.  It is not intended for use as a debugging tool.

Fixes EUCA-9377