Instance launch/terminate updated for VPC - EUCA-9605
The EC2 service now supports basic launching / termination of instances
in a VPC.

A default security group is now created for a VPC and will be deleted
when the VPC is deleted. On instance launch the default VPC security
group will be used if a group is not specified. The availability zone
is determined from the subnet and a private IP for the subnet must be
specified. A network interface is created for the instance and will be
deleted when the instance is terminated.

EUCA-9605